requirements.txt
numpy
pandas
pandas_stubs
Pillow
requests
streamlit
tensorflow
